many changes

+++ pacbio_experiment.py
@@ -11,12 +11,15 @@ from collections import defaultdict, Counter
 
 import pandas as pd
 import matplotlib.pyplot as plt
+import numpy as np
 import bokeh
 import pysam
 import yaml
 
+import Sequencing.sam as sam
 import Sequencing.fastq as fastq
 import Sequencing.utilities as utilities
+import Sequencing.visualize_structure as visualize_structure
 
 import target_info
 import blast
@@ -31,11 +34,12 @@ for i, source in enumerate(['PCR', 'plasmid', 'ssDNA', 'CT']):
     for replicate in [1, 2, 3]:
         source_to_color[source, replicate] = palette[4 * i  + (replicate - 1)]
 
-def priority_tuple_to_string(t):
-    return '{0}: {1}'.format(*t)
-
-def priority_string_to_description(s):
-    return s.split(': ')[1]
+palette = bokeh.palettes.Set2[8]
+cap_to_color = {
+    'AmMC6': palette[0],
+    'Biotin': palette[1],
+    'IDDT': palette[2],
+}
 
 class PacbioExperiment(object):
     def __init__(self, name):
@@ -52,25 +56,39 @@ class PacbioExperiment(object):
             'bam_by_name': self.dir / 'alignments.by_name.bam',
             'outcomes_dir': self.dir / 'outcomes',
             'outcome_counts': self.dir / 'outcome_counts.csv',
+            'outcome_sort_order': self.dir / 'outcome_sort_order.txt',
             'lengths_figure': self.dir / 'all_lengths.png',
+            'manual_length_ranges': self.dir / 'manual_length_ranges.csv',
         }
 
         self.cell_line = description.get('cell_line')
         self.donor_type = description.get('donor_type')
-        self.replicate = description.get('replicate')
+        self.capped = description.get('capped')
+        self.cap = description.get('cap')
+        self.replicate = description.get('replicate', 1)
         self.sorted = description.get('sorted')
 
         self.color = source_to_color.get((self.donor_type, self.replicate), 'grey')
+        if self.capped:
+            self.color = cap_to_color[self.cap]
+    
+    def outcome_fns(self, outcome):
+        outcome_string = '_'.join(map(str, outcome))
+        outcome_dir = self.fns['outcomes_dir'] / outcome_string
+        fns = {
+            'dir': outcome_dir,
+            'query_names': outcome_dir / 'qnames.txt',
+            'bam_by_name': outcome_dir / 'alignments.by_name.bam',
+            'first_example': outcome_dir / 'first_examples.png',
+            'combined_figure': outcome_dir / 'combined.png',
+            'lengths_figure': outcome_dir / 'lengths.png',
+            'text_alignments': outcome_dir / 'alignments.txt',
+        }
+        return fns
 
     def query_names(self):
-        for read in fastq.reads(self.fns['fastq']):
+        for read in fastq.reads(self.fns['fastq'], up_to_space=True):
             yield read.name
-
-    def outcome_query_names(self, outcome):
-        outcome_fns = self.outcome_fns(outcome)
-        als = pysam.AlignmentFile(outcome_fns['bam_by_name'])
-        names = [n for n, _ in utilities.group_by(als, lambda al: al.query_name)]
-        return names
 
     @utilities.memoized_property
     def read_lengths(self):
@@ -78,12 +96,20 @@ class PacbioExperiment(object):
         lengths = utilities.counts_to_array(lengths)
         return lengths
 
+    @property
+    def length_ranges(self):
+        path = self.fns['manual_length_ranges']
+        if path.exists():
+            ranges = pd.read_csv(path, sep='\t', header=None, names=['start', 'end'])
+        else:
+            ranges = pd.DataFrame(columns=['start', 'end'])
+        return ranges
+
     def outcome_read_lengths(self, outcome):
         outcome_fns = self.outcome_fns(outcome)
-        als = pysam.AlignmentFile(outcome_fns['bam_by_name'])
 
         lengths = Counter()
-        for _, group in utilities.group_by(als, lambda al: al.query_name):
+        for _, group in sam.grouped_by_name(outcome_fns['bam_by_name']):
             lengths[group[0].query_length] += 1
 
         lengths = utilities.counts_to_array(lengths)
@@ -117,43 +143,51 @@ class PacbioExperiment(object):
         else:
             return list(counts.index)
 
-    def outcome_fns(self, outcome):
-        outcome_string = '_'.join(map(str, outcome))
-        outcome_dir = self.fns['outcomes_dir'] / outcome_string
-        fns = {
-            'dir': outcome_dir,
-            'query_names': outcome_dir / 'qnames.txt',
-            'bam_by_name': outcome_dir / 'alignments.by_name.bam',
-            'first_example': outcome_dir / 'first_examples.png',
-            'combined_figure': outcome_dir / 'combined.png',
-            'lengths_figure': outcome_dir / 'lengths.png',
-            'text_alignments': outcome_dir / 'alignments.txt',
-        }
-        return fns
-
     def outcome_query_names(self, outcome):
         fns = self.outcome_fns(outcome)
         qnames = [l.strip() for l in open(fns['query_names'])]
         return qnames
+    
+    def load_outcome_sort_order(self):
+        sort_order = {}
+        for line in self.fns['outcome_sort_order'].open():
+            outcome_string, priority_string = line.strip().split('\t')
+            outcome = tuple(outcome_string.split('_'))
+            priority = tuple(int(p) for p in priority_string.split('_'))
+            sort_order[outcome] = priority
+        return sort_order
 
     def count_outcomes(self):
+        if self.fns['outcomes_dir'].is_dir():
+            shutil.rmtree(self.fns['outcomes_dir'])
+        self.fns['outcomes_dir'].mkdir()
+
         bam_fh = pysam.AlignmentFile(self.fns['bam_by_name'])
-        alignment_groups = utilities.group_by(bam_fh, lambda al: al.query_name)
+        alignment_groups = sam.grouped_by_name(bam_fh)
         outcomes = defaultdict(list)
 
+        sort_order = {}
         for name, als in alignment_groups:
             layout_info = layout.characterize_layout(als, self.target_info)
-            outcomes[layout_info['outcome']].append(name)
+            
+            outcome = layout_info['outcome']
+            outcomes[outcome['description']].append(name)
+
+            sort_order[outcome['description']] = outcome['sort_order']
 
         bam_fh.close()
-
-        shutil.rmtree(self.fns['outcomes_dir'])
-        self.fns['outcomes_dir'].mkdir()
 
         counts = {description: len(names) for description, names in outcomes.items()}
         series = pd.Series(counts)
 
         series.to_csv(self.fns['outcome_counts'], sep='\t')
+
+        with open(self.fns['outcome_sort_order'], 'w') as fh:
+            for outcome, priority in sort_order.items():
+                outcome_string = '_'.join(outcome)
+                priority_string = '_'.join(map(str, priority))
+
+                fh.write('{0}\t{1}\n'.format(outcome_string, priority_string))
         
         # To make plotting easier, for each outcome, make a file listing all of
         # qnames for the outcome and a bam file (sorted by name) with all of the
@@ -183,46 +217,104 @@ class PacbioExperiment(object):
             fh.close()
 
     def make_outcome_plots(self, num_examples=10):
-        fig = visualize.make_length_plot(self.read_lengths, self.color)
+        fig = self.length_distribution_figure()
         fig.savefig(str(self.fns['lengths_figure']), bbox_inches='tight')
         plt.close(fig)
 
         for outcome in self.outcomes:
             outcome_fns = self.outcome_fns(outcome)
             
-            with tempfile.TemporaryDirectory(suffix='_outcome_plots') as temp_dir:
-                temp_fns = []
-                for i in range(num_examples):
-                    fig = visualize.plot_read(self.name, i, outcome=outcome, parsimonious=True)
-                    if fig is None:
-                        continue
-                        
-                    fig.axes[0].set_title('_', y=1.2, color='white')
-                    
-                    fn = Path(temp_dir) / '{0:05d}.png'.format(i)
-                    temp_fns.append(fn)
-                    fig.savefig(str(fn), bbox_inches='tight')
-                    
-                    if i == 0:
-                        fig.axes[0].set_title('')
-                        fig.savefig(str(outcome_fns['first_example']), bbox_inches='tight')
-                        
-                    plt.close(fig)
+            als = self.get_read_alignments(0, outcome)
+            fig = visualize.plot_read(als, self.target_info, parsimonious=True)
+            fig.axes[0].set_title('')
+            fig.savefig(str(outcome_fns['first_example']), bbox_inches='tight')
+            plt.close(fig)
+            
+            als_iter = (self.get_read_alignments(i, outcome) for i in range(num_examples))
+            stacked_im = visualize.make_stacked_Image(als_iter, self.target_info, parsimonious=True)
+            stacked_im.save(outcome_fns['combined_figure'])
 
-                lengths = self.outcome_read_lengths(outcome)
-
-                fig = visualize.make_length_plot(self.read_lengths, self.color, lengths)
-                fig.savefig(str(outcome_fns['lengths_figure']), bbox_inches='tight')
-                plt.close(fig)
+            lengths = self.outcome_read_lengths(outcome)
+            fig = visualize.make_length_plot(self.read_lengths, self.color, lengths)
+            fig.savefig(str(outcome_fns['lengths_figure']), bbox_inches='tight')
+            plt.close(fig)
                 
-                to_concat = [outcome_fns['lengths_figure']] + temp_fns                
-                convert_command = ['convert'] + to_concat + ['-background', 'white', '-gravity', 'center', '-append', outcome_fns['combined_figure']]
-                subprocess.check_call(convert_command)
-    
+    def get_read_alignments(self, read_id, outcome=None):
+        if outcome is not None:
+            bam_fn = self.outcome_fns(outcome)['bam_by_name']
+        else:
+            bam_fn = self.fns['bam_by_name']
+        
+        read_groups = sam.grouped_by_name(bam_fn)
+
+        if isinstance(read_id, int):
+            try:
+                for i in range(read_id + 1):
+                    name, group = next(read_groups)
+                return group
+            except StopIteration:
+                return None
+        else:
+            for name, group in read_groups:
+                if name == read_id:
+                    break
+            if name == read_id:
+                return group
+            else:
+                return None
+
+    def make_text_visualizations(self, num_examples=10):
+        for outcome in self.outcomes:
+            outcome_fns = self.outcome_fns(outcome)
+            visualize_structure.visualize_bam_alignments(outcome_fns['bam_by_name'],
+                                                         self.target_info.fns['ref_fasta'],
+                                                         outcome_fns['text_alignments'],
+                                                         num_examples,
+                                                        )
+    def length_distribution_figure(self, show_ranges=False):
+        ys = self.read_lengths
+
+        fig, ax = plt.subplots(figsize=(16, 5))
+
+        ax.plot(ys, color=self.color)
+        ax.set_ylim(0, 1.01 * max(ys))
+        ax.set_xlim(0, len(ys))
+                           
+        if show_ranges:
+            for _, (start, end) in self.length_ranges.iterrows():
+                ax.axvspan(start, end,
+                           gid='length_range_{0:05d}_{1:05d}'.format(start, end),
+                           alpha=0.1,
+                           facecolor='white',
+                           edgecolor='black',
+                           zorder=100,
+                          )
+            
+        major = np.arange(0, len(ys), 500)
+        minor = [x for x in np.arange(0, len(ys), 100) if x % 500 != 0]
+                    
+        ax.set_xticks(major)
+        ax.set_xticks(minor, minor=True)
+
+        ax.set_ylabel('Number of reads')
+        ax.set_xlabel('Length of read')
+
+        return fig
+
+    def span_to_Image(self, start, end, num_examples=5):
+        groups = sam.grouped_by_name(self.fns['bam_by_name'])
+        filtered = (group for name, group in groups
+                    if start <= group[0].query_length <= end)
+
+        sample = utilities.reservoir_sample(filtered, num_examples)
+        
+        return visualize.make_stacked_Image(sample, self.target_info, parsimonious=True)
+        
     def process(self):
-        #self.generate_alignments()
-        #self.count_outcomes()
-        self.make_outcome_plots()
+        self.generate_alignments()
+        self.count_outcomes()
+        #self.make_outcome_plots(num_examples=5)
+        #self.make_text_visualizations()
 
 def get_all_experiments(conditions=None):
     if conditions is None:
@@ -255,7 +347,9 @@ if __name__ == '__main__':
     if args.parallel is not None:
         max_procs = args.parallel
 
-        exps = get_all_experiments()
+        #conditions = {'target_name': 'CLTA-150HA'}
+        conditions = {}
+        exps = get_all_experiments(conditions)
         names = sorted(exp.name for exp in exps)
 
         parallel_command = [
